==> pages/05_Future_work.py <==
import streamlit as st



st.title('Future Work')

'''
1. Change heatmap to one combined image. This hasn't been performed as the grid coordinates would be difficult to work out without creating a lot of overlap between main holds and aux holds
2. Fix dependency issues stopping the random forest and LSTM models to load
3. Fix create new climbs so that when angle changes, grade remains the same

'''
